+++ .python-version
@@ -1,1 +1,1 @@
-3.10.0
+3.11.11
