+++ EvolverExample.json
@@ -183,17 +183,17 @@
         "completed_at": null
     },
     "config": {
-        "num_generations": 100,
+        "num_generations": 10,
         "parallel_processing": 4,
         "use_parallel": true,
-        "sol_per_pop": 500,
-        "num_parents_mating": 20,
-        "mutation_percent_genes": [],
+        "sol_per_pop": 5,
+        "num_parents_mating": 2,
+        "mutation_percent_genes": 20,
         "crossover_type": "two_points",
         "mutation_type": "adaptive",
         "parent_selection_type": "rank",
-        "K_tournament": 10,
-        "elitism": 5,
+        "K_tournament": 3,
+        "elitism": 2,
         "diversity_threshold": 0.6,
         "stagnation_threshold": 20,
         "early_stopping_patience": 50,
