+++ fast_trade/ml/evolution/genetic_algorithm.py
@@ -3,310 +3,19 @@ import datetime
 import json
 import os
 import random
-import uuid
-from dataclasses import dataclass, field
 from functools import partial
-from typing import Any, Callable, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 
 import numpy as np
 import requests
 
 from fast_trade import run_backtest
+from fast_trade.ml.evolution.models import GeneDefinition, OptimizationConfig, OptimizationResult
+from fast_trade.ml.evolution.strategy_modifier import modify_strategy
+from fast_trade.ml.evolution.utils import evaluate_solution_wrapper
 
 # Constants
-FREQUENCY_MAP = ["1Min", "5Min", "15Min", "30Min", "1h", "4h", "8h", "12h"]
-COLUMNS = ["close", "open", "low", "high"]
 ARCHIVE_PATH = os.getenv("ARCHIVE_PATH", os.path.join(os.getcwd(), "ft_archive/ml"))
-
-
-@dataclass
-class GeneDefinition:
-    """Definition of a gene in the genetic algorithm."""
-
-    name: str
-    type: str  # 'int', 'float', 'categorical', 'boolean'
-    min_value: float = 0.0
-    max_value: float = 100.0
-    categories: List[str] = field(default_factory=list)
-    constraints: List[Dict[str, Any]] = field(default_factory=list)
-
-
-@dataclass
-class OptimizationConfig:
-    """Configuration for genetic algorithm optimization."""
-
-    num_generations: int = 100
-    num_parents_mating: int = 10
-    sol_per_pop: int = 20
-    parent_selection_type: str = "tournament"
-    crossover_type: str = "uniform"
-    mutation_type: str = "adaptive"
-    mutation_percent_genes: float = field(default=0.1, metadata={"type": float})
-    parallel_processing: int = 8
-    use_parallel: bool = False  # Flag to enable/disable parallel processing
-    K_tournament: int = 4
-    elitism: int = 2  # Number of best solutions to preserve
-    diversity_threshold: float = 0.7
-    stagnation_threshold: int = 10
-    early_stopping_patience: int = 20
-    min_improvement: float = 0.001
-    fitness: Dict[str, Any] = field(default_factory=dict)
-    refresh_generations: int = 10
-    refresh_percent: float = 0.5
-
-    def __post_init__(self):
-        """Ensure mutation_percent_genes is a float."""
-        self.mutation_percent_genes = float(self.mutation_percent_genes)
-
-
-@dataclass
-class OptimizationResult:
-    """Results from the optimization process."""
-
-    mapped_genes: List[Tuple[str, str]]
-    fitness: float
-    best_strategy: Dict[str, Any]
-    started_at: datetime.datetime
-    completed_at: datetime.datetime
-    generation_history: List[Dict[str, Any]]
-
-
-def modify_strategy(
-    strategy: Dict[str, Any], genes: List[Tuple[str, str]]
-) -> Dict[str, Any]:
-    """Create a new strategy with evolved genes."""
-    # Create a new strategy object with predefined sets
-    new_strategy = {
-        "predefined_sets": {
-            "columns": COLUMNS,
-            "operators": ["<", ">", "=", "!="],
-            "transformers": ["ema", "zlema", "sma"],
-            "frequencies": FREQUENCY_MAP,
-        }
-    }
-
-    # Create a mapping of gene names to values
-    gene_map = {gene_name: gene_value for gene_name, gene_value in genes}
-
-    # Set frequency
-    if "freq" in gene_map:
-        freq_idx = int(float(gene_map["freq"]) * (len(FREQUENCY_MAP) - 1))
-        freq_idx = max(0, min(freq_idx, len(FREQUENCY_MAP) - 1))
-        new_strategy["freq"] = FREQUENCY_MAP[freq_idx]
-    else:
-        new_strategy["freq"] = strategy.get("freq", "1h")
-
-    # Create datapoints
-    new_strategy["datapoints"] = []
-    for datapoint in strategy.get("datapoints", []):
-        new_datapoint = datapoint.copy()
-        base_name = new_datapoint["name"]
-
-        # Handle transformer
-        if f"{base_name}_transformer" in gene_map:
-            try:
-                float_value = float(gene_map[f"{base_name}_transformer"])
-                transformer_idx = int(
-                    float_value
-                    * (len(new_strategy["predefined_sets"]["transformers"]) - 1)
-                )
-                transformer_idx = max(
-                    0,
-                    min(
-                        transformer_idx,
-                        len(new_strategy["predefined_sets"]["transformers"]) - 1,
-                    ),
-                )
-                new_datapoint["transformer"] = new_strategy["predefined_sets"][
-                    "transformers"
-                ][transformer_idx]
-            except (ValueError, TypeError):
-                new_datapoint["transformer"] = "sma"
-
-        # Handle args (periods)
-        if "args" in new_datapoint:
-            new_args = []
-            for arg in new_datapoint["args"]:
-                if isinstance(arg, str) and arg.startswith("#"):
-                    gene_name = arg[1:]  # Remove the # prefix
-                    if gene_name in gene_map:
-                        # Try to convert string to int if it represents a numeric value
-                        try:
-                            new_args.append(int(float(gene_map[gene_name])))
-                        except (ValueError, TypeError):
-                            new_args.append(gene_map[gene_name])
-                    else:
-                        new_args.append(arg)
-                else:
-                    new_args.append(arg)
-            new_datapoint["args"] = new_args
-
-        new_strategy["datapoints"].append(new_datapoint)
-
-    # Create enter conditions
-    new_strategy["enter"] = []
-    for condition in strategy.get("enter", []):
-        new_condition = condition.copy()
-        base_name = new_condition[0]
-
-        # Handle operator
-        if f"{base_name}_operator" in gene_map:
-            try:
-                float_value = float(gene_map[f"{base_name}_operator"])
-                operator_idx = int(
-                    float_value
-                    * (len(new_strategy["predefined_sets"]["operators"]) - 1)
-                )
-                operator_idx = max(
-                    0,
-                    min(
-                        operator_idx,
-                        len(new_strategy["predefined_sets"]["operators"]) - 1,
-                    ),
-                )
-                new_condition[1] = new_strategy["predefined_sets"]["operators"][
-                    operator_idx
-                ]
-            except (ValueError, TypeError):
-                new_condition[1] = ">"
-
-        # Handle column
-        if f"{base_name}_column" in gene_map:
-            try:
-                float_value = float(gene_map[f"{base_name}_column"])
-                column_idx = int(
-                    float_value * (len(new_strategy["predefined_sets"]["columns"]) - 1)
-                )
-                column_idx = max(
-                    0,
-                    min(
-                        column_idx, len(new_strategy["predefined_sets"]["columns"]) - 1
-                    ),
-                )
-                new_condition[2] = new_strategy["predefined_sets"]["columns"][
-                    column_idx
-                ]
-            except (ValueError, TypeError):
-                new_condition[2] = "close"
-        elif isinstance(new_condition[2], str) and new_condition[2].startswith("#"):
-            gene_name = new_condition[2][1:]  # Remove the # prefix
-            if gene_name in gene_map:
-                # Try to convert to int if it represents a numeric value
-                try:
-                    new_condition[2] = int(float(gene_map[gene_name]))
-                except (ValueError, TypeError):
-                    new_condition[2] = gene_map[gene_name]
-
-        new_strategy["enter"].append(new_condition)
-
-    # Create exit conditions (similar to enter)
-    new_strategy["exit"] = []
-    for condition in strategy.get("exit", []):
-        new_condition = condition.copy()
-        base_name = new_condition[0]
-
-        # Handle operator
-        if f"{base_name}_operator" in gene_map:
-            try:
-                float_value = float(gene_map[f"{base_name}_operator"])
-                operator_idx = int(
-                    float_value
-                    * (len(new_strategy["predefined_sets"]["operators"]) - 1)
-                )
-                operator_idx = max(
-                    0,
-                    min(
-                        operator_idx,
-                        len(new_strategy["predefined_sets"]["operators"]) - 1,
-                    ),
-                )
-                new_condition[1] = new_strategy["predefined_sets"]["operators"][
-                    operator_idx
-                ]
-            except (ValueError, TypeError):
-                new_condition[1] = ">"
-
-        # Handle column
-        if f"{base_name}_column" in gene_map:
-            try:
-                float_value = float(gene_map[f"{base_name}_column"])
-                column_idx = int(
-                    float_value * (len(new_strategy["predefined_sets"]["columns"]) - 1)
-                )
-                column_idx = max(
-                    0,
-                    min(
-                        column_idx, len(new_strategy["predefined_sets"]["columns"]) - 1
-                    ),
-                )
-                new_condition[2] = new_strategy["predefined_sets"]["columns"][
-                    column_idx
-                ]
-            except (ValueError, TypeError):
-                new_condition[2] = "close"
-        elif isinstance(new_condition[2], str) and new_condition[2].startswith("#"):
-            gene_name = new_condition[2][1:]  # Remove the # prefix
-            if gene_name in gene_map:
-                # Try to convert to int if it represents a numeric value
-                try:
-                    new_condition[2] = int(float(gene_map[gene_name]))
-                except (ValueError, TypeError):
-                    new_condition[2] = gene_map[gene_name]
-
-        new_strategy["exit"].append(new_condition)
-
-    # Copy other strategy properties
-    for key, value in strategy.items():
-        if key not in ["predefined_sets", "freq", "datapoints", "enter", "exit"]:
-            new_strategy[key] = value
-
-    return new_strategy
-
-
-def save_optimization_results(result: OptimizationResult, run_id: str) -> None:
-    """Save optimization results to a file."""
-    results_dir = f"{ARCHIVE_PATH}/{run_id}"
-    os.makedirs(results_dir, exist_ok=True)
-
-    result_dict = {
-        "mapped_genes": result.mapped_genes,
-        "fitness": result.fitness,
-        "best_strategy": result.best_strategy,
-        "started_at": result.started_at.isoformat(),
-        "completed_at": result.completed_at.isoformat(),
-        "generation_history": result.generation_history,
-    }
-
-    filename = f"{results_dir}/winner.json"
-    with open(filename, "w", encoding="utf-8") as f:
-        json.dump(result_dict, f, indent=2)
-
-
-def evaluate_solution_wrapper(solution, base_strategy, fitness_weights):
-    """Standalone wrapper function for parallel evaluation of solutions.
-    This must be at module level to be picklable for multiprocessing."""
-    # Convert solution to list of tuples for modify_strategy
-    solution_tuples = [(k, str(v)) for k, v in solution.items()]
-    strategy = modify_strategy(base_strategy.copy(), solution_tuples)
-    result = run_backtest(strategy)
-
-    # Calculate fitness using multiple metrics
-    metrics = result.get("summary", {})
-    fitness = 0
-
-    def get_metric_value(metric):
-        keys = metric.split(".")
-        value = metrics
-        for key in keys:
-            value = value.get(key, 0)  # Default to 0 if the key is not found
-        return value
-
-    for metric, weight in fitness_weights.items():
-        metric_value = get_metric_value(metric)
-        # make sure to keep the metric value as it is, dont convert let it be positive
-        fitness += metric_value * weight
-
-    return (fitness, metrics)
 
 
 class GeneticAlgorithm:
@@ -658,7 +367,7 @@ class GeneticAlgorithm:
             new_population: List[Dict[str, Any]] = []
 
             # Elitism: preserve best solutions
-            elite_indices = np.argsort(self.fitness_scores)[-self.config.elitism :]
+            elite_indices = np.argsort(self.fitness_scores)[-self.config.elitism:]
             new_population.extend([self.population[i] for i in elite_indices])
 
             # Create offspring
@@ -734,217 +443,4 @@ class GeneticAlgorithm:
             started_at=self.started_at,
             completed_at=datetime.datetime.now(),
             generation_history=self.generation_history,
-        )
-
-
-def optimize_strategy(
-    base_strategy: Dict[str, Any],
-    genes: List[Tuple[str, Any]],
-    config: Optional[OptimizationConfig] = None,
-    run_id: str = "default",
-    config_file: Dict = {},
-    api_url: Optional[str] = None,
-) -> OptimizationResult:
-    """
-    Optimizes a trading strategy using a genetic algorithm.
-
-    Args:
-        base_strategy: A dictionary representing the base trading strategy.
-        genes: A list of tuples containing (gene_name, gene_value).
-        config: Optional configuration for the genetic algorithm.
-
-    Returns:
-        OptimizationResult containing the best solution and metadata.
-    """
-    # save all the inputs to a json file with the run_id
-    # make the directory if it doesn't exist
-    os.makedirs(f"{ARCHIVE_PATH}/{run_id}", exist_ok=True)
-    with open(f"{ARCHIVE_PATH}/{run_id}/inputs.json", "w", encoding="utf-8") as f:
-        json.dump(
-            {
-                "run_id": run_id,
-                **config_file,
-            },
-            f,
-            indent=2,
-        )
-        
-    if api_url:
-        print("Reporting to: ", api_url)
-
-    config = config or OptimizationConfig()
-
-    genes = process_genes_from_config(genes)
-    # Convert gene tuples to GeneDefinition objects
-    gene_definitions: List[GeneDefinition] = []
-    for gene_name, gene_value in genes:
-        if callable(gene_value):
-            # For callable genes, determine type and range
-            sample_value = gene_value()
-            if isinstance(sample_value, int):
-                # Special handling for frequency gene
-                if gene_name == "freq":
-                    gene_definitions.append(
-                        GeneDefinition(
-                            name=gene_name,
-                            type="int",
-                            min_value=0.0,
-                            max_value=float(len(FREQUENCY_MAP) - 1),
-                        )
-                    )
-                # Special handling for period genes
-                elif gene_name.endswith("_period"):
-                    gene_definitions.append(
-                        GeneDefinition(
-                            name=gene_name,
-                            type="int",
-                            min_value=1.0,  # Ensure minimum period is 1
-                            max_value=100.0,
-                        )
-                    )
-                else:
-                    gene_definitions.append(
-                        GeneDefinition(
-                            name=gene_name, type="int", min_value=0.0, max_value=100.0
-                        )
-                    )
-            else:
-                # For column selection, use float type to allow continuous values
-                if gene_name.endswith("_column"):
-                    gene_definitions.append(
-                        GeneDefinition(
-                            name=gene_name, type="float", min_value=0.0, max_value=1.0
-                        )
-                    )
-                else:
-                    gene_definitions.append(
-                        GeneDefinition(
-                            name=gene_name, type="float", min_value=-1.0, max_value=1.0
-                        )
-                    )
-        else:
-            # For static genes, determine type
-            if isinstance(gene_value, bool):
-                gene_definitions.append(GeneDefinition(name=gene_name, type="boolean"))
-            elif isinstance(gene_value, str):
-                gene_definitions.append(
-                    GeneDefinition(
-                        name=gene_name, type="categorical", categories=[gene_value]
-                    )
-                )
-            else:
-                gene_definitions.append(
-                    GeneDefinition(
-                        name=gene_name, type="float", min_value=-1.0, max_value=1.0
-                    )
-                )
-
-    # Create and run genetic algorithm
-    ga = GeneticAlgorithm(
-        base_strategy,
-        gene_definitions,
-        config,
-        fitness=config.fitness,
-        run_id=run_id,
-        api_url=api_url,
-    )
-    result = ga.run()
-
-    # Save results
-    save_optimization_results(result, run_id)
-
-    return result
-
-
-def process_genes_from_config(input_genes: List[Dict]) -> List[Tuple[str, Any]]:
-    """Process genes from config by creating appropriate generator functions."""
-    new_genes = []
-
-    # Define gene generator functions
-    def create_int_generator(min_val: int, max_val: int) -> Callable:
-        return lambda: random.randint(min_val, max_val)
-
-    def create_float_generator(min_val: float, max_val: float) -> Callable:
-        return lambda: random.uniform(min_val, max_val)
-
-    def create_categorical_generator(values: List[str], gene_name: str) -> Callable:
-        if not values:
-            raise ValueError(
-                f"Cannot create categorical generator for gene '{gene_name}' with empty values list"
-            )
-        return lambda: random.choice(values)
-
-    for gene in input_genes:
-        gene_type = gene.get("type", "int")
-        gene_name = gene["name"]
-        values_ref = gene.get("values_ref", None)
-
-        if gene_type == "int":
-            range_values = gene.get("range", [0, 100])
-            new_genes.append(
-                (gene_name, create_int_generator(range_values[0], range_values[1]))
-            )
-        elif gene_type == "float":
-            range_values = gene.get("range", [0.0, 100.0])
-            new_genes.append(
-                (gene_name, create_float_generator(range_values[0], range_values[1]))
-            )
-        elif gene_type == "categorical":
-            if values_ref == "columns":
-                values = COLUMNS
-            elif values_ref == "operators":
-                values = ["<", ">", "=", "!="]  # Default operators if not specified
-            elif values_ref == "transformers":
-                values = [
-                    "ema",
-                    "zlema",
-                    "sma",
-                ]  # Default transformers if not specified
-            elif values_ref == "frequencies":
-                values = FREQUENCY_MAP
-            else:
-                values = values_ref if isinstance(values_ref, list) else []
-
-            try:
-                new_genes.append(
-                    (gene_name, create_categorical_generator(values, gene_name))
-                )
-            except ValueError as e:
-                print(f"Error processing gene {gene_name}: {str(e)}")
-                # Fallback to a default value based on the gene name
-                if "operator" in gene_name:
-                    values = ["<", ">", "=", "!="]
-                elif "transformer" in gene_name:
-                    values = ["ema", "zlema", "sma"]
-                elif "column" in gene_name:
-                    values = COLUMNS
-                else:
-                    values = ["default"]  # Fallback for unknown categorical genes
-                new_genes.append(
-                    (gene_name, create_categorical_generator(values, gene_name))
-                )
-        else:
-            # Default to int generator if type is unknown
-            new_genes.append((gene_name, create_int_generator(0, 100)))
-
-    return new_genes
-
-
-def run_evolver(evolver_config: Dict):
-    """ Run the evolver with the given config """
-    base_strategy = evolver_config["strategy"]
-    genes = evolver_config["genes"]
-
-    config = OptimizationConfig(
-        **evolver_config["config"], fitness=evolver_config["fitness"]
-    )
-    run_id = str(uuid.uuid4())
-
-    optimize_strategy(
-        base_strategy=base_strategy,
-        genes=genes,
-        config=config,
-        run_id=run_id,
-        config_file=evolver_config,
-        api_url=evolver_config.get("api_url", None),
-    )
+        )
