+++ fast_trade/ml/evolution/models.py
@@ -1,0 +1,55 @@
+import datetime
+from dataclasses import dataclass, field
+from typing import Any, Dict, List
+
+
+@dataclass
+class GeneDefinition:
+    """Definition of a gene in the genetic algorithm."""
+
+    name: str
+    type: str  # 'int', 'float', 'categorical', 'boolean'
+    min_value: float = 0.0
+    max_value: float = 100.0
+    categories: List[str] = field(default_factory=list)
+    constraints: List[Dict[str, Any]] = field(default_factory=list)
+
+
+@dataclass
+class OptimizationConfig:
+    """Configuration for genetic algorithm optimization."""
+
+    num_generations: int = 100
+    num_parents_mating: int = 10
+    sol_per_pop: int = 20
+    parent_selection_type: str = "tournament"
+    crossover_type: str = "uniform"
+    mutation_type: str = "adaptive"
+    mutation_percent_genes: float = field(default=0.1, metadata={"type": float})
+    parallel_processing: int = 8
+    use_parallel: bool = False  # Flag to enable/disable parallel processing
+    K_tournament: int = 4
+    elitism: int = 2  # Number of best solutions to preserve
+    diversity_threshold: float = 0.7
+    stagnation_threshold: int = 10
+    early_stopping_patience: int = 20
+    min_improvement: float = 0.001
+    fitness: Dict[str, Any] = field(default_factory=dict)
+    refresh_generations: int = 10
+    refresh_percent: float = 0.5
+
+    def __post_init__(self):
+        """Ensure mutation_percent_genes is a float."""
+        self.mutation_percent_genes = float(self.mutation_percent_genes)
+
+
+@dataclass
+class OptimizationResult:
+    """Results from the optimization process."""
+
+    mapped_genes: List[tuple[str, str]]
+    fitness: float
+    best_strategy: Dict[str, Any]
+    started_at: datetime.datetime
+    completed_at: datetime.datetime
+    generation_history: List[Dict[str, Any]]
